servers/document-extraction-mcp/pyproject.toml
[project]
name = "document-extraction-mcp"
version = "0.1.0"
description = "MCP Server for document extraction (OCR, Layout Detection, Reading Order, VLM Analysis)"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # MCP Server
    "mcp>=1.0.0",

    # Image Processing
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",

    # OCR + Layout Detection
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",

    # Reading Order (LayoutLMv3)
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "layoutreader @ git+https://github.com/ppaanngggg/layoutreader.git",

    # VLM calls
    "litellm>=1.0.0",
    "aiohttp>=3.9.0",

    # Utilities
    "structlog>=24.0.0",
]

[project.scripts]
document-extraction-mcp = "document_extraction_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/document_extraction_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
